refactor(server): split proxy logic into core, controller and route modules
