+++ projects/server/src/config/serverConfig.ts
@@ -1,0 +1,3 @@
+export const PORT = 3000;
+export const HOST = 'localhost';
+export const BASE_URL = `http://${HOST}:${PORT}`;
