+++ projects/server/src/types/IProxy.ts
@@ -1,0 +1,12 @@
+import { Readable } from 'node:stream';
+
+export interface IProxyResult {
+  headers: { [key: string]: string };
+  body?: unknown;
+  stream?: Readable;
+  status: number;
+}
+
+export default interface IProxy {
+  execute(destinationUrl: string): Promise<IProxyResult>;
+}
